requirements.txt
pytest
pylint
ipdb
jupyterlab
numpy
pandas
matplotlib
seaborn
scikit-learn
db-dtypes
google-cloud-bigquery
google-cloud-bigquery-storage
google-cloud-storage



fastapi
uvicorn
openai
pillow
python-dotenv
python-multipart
tensorflow
joblib
